style: clear Trending code
